Type Definition rusoto_credential::DefaultCredentialsProvider

type DefaultCredentialsProvider = AutoRefreshingProvider<ChainProvider>;

The credentials provider you probably want to use if you don't require Sync for your AWS services. Wraps a ChainProvider in an AutoRefreshingProvider that uses a RefCell to cache credentials

The underlying ChainProvider checks multiple sources for credentials, and the AutoRefreshingProvider refreshes the credentials automatically when they expire. The RefCell allows this caching to happen without the overhead of a Mutex, but is !Sync.

For a Sync implementation of the same, see DefaultCredentialsProviderSync