Crate rusoto_credential

Types for loading and managing AWS access credentials for API requests.

Modules

claims
Credential Claims module.

Structs

AutoRefreshingProvider
Wrapper for ProvideAwsCredentials that caches the credentials returned by the wrapped provider. Each time the credentials are accessed, they are checked to see if they have expired, in which case they are retrieved from the wrapped provider again.

AutoRefreshingProviderFuture
Future returned from AutoRefreshingProvider.

AwsCredentials
AWS API access credentials, including access key, secret key, token (for IAM profiles), expiration timestamp, and claims from federated login.

ChainProvider
Provides AWS credentials from multiple possible sources using a priority order.

ChainProviderFuture
Future returned from ChainProvider.

ContainerProvider
Provides AWS credentials from a task's IAM role.

ContainerProviderFuture
Future returned from ContainerProvider.

CredentialsError
Represents an Error that has occured during the fetching Credentials Phase.

DefaultCredentialsProvider
Wraps a ChainProvider in an AutoRefreshingProvider.

DefaultCredentialsProviderFuture
Future returned from DefaultCredentialsProvider.

EnvironmentProvider
Provides AWS credentials from environment variables.

EnvironmentProviderFuture
Provides AWS credentials from environment variables as a Future.

InstanceMetadataProvider
Provides AWS credentials from a resource's IAM role.

InstanceMetadataProviderFuture
Future returned from InstanceMetadataProvider.

ProfileProvider
Provides AWS credentials from a profile in a credentials file, or from a credential process.

ProfileProviderFuture
Provides AWS credentials from a profile in a credentials file as a Future.

StaticProvider
Provides AWS credentials from statically/programmatically provided strings.

Traits

ProvideAwsCredentials
A trait for types that produce AwsCredentials.